Move PrimeFaces Selenium testers into primefaces package
