+++ test/integration/src/test/java/com/liferay/faces/bridge/test/integration/issue/primefaces/FACES_3031PortletTester.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2017 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-2018 Liferay, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.liferay.faces.bridge.test.integration.issue;
+package com.liferay.faces.bridge.test.integration.issue.primefaces;
 
 import org.junit.Test;
+
+import com.liferay.faces.bridge.test.integration.issue.SimpleFACESPortletTester;
 
 
 /**
